Proposed fix for null dutyDefToSubNode. I don't really get WHY it works I just know that after a LOT of troubleshooting and following code flow I ended up finding the problem was XML related, popping up when ThinkTreeDefs/Humanlike.xml loaded successfully from Combat Realism.
Some more troubleshooting led me to LordDutyConstant near the end of the file that when present the error happened.  Checked Core and it's there too.  Noticed it was defined in SubTrees_Duty.xml so just copied that file over from Core into Combat Realism.  No errors.

The part that confuses me is that I thought once something was defined in a lower level and non-abstract (ie the treeDef LordDutyConstant in Core) that you didn't need to re-define it higher up (ie Combat Realism).